Made firebase mocking it's own file, made a helper file with functions to use, split up admins.test.ts into a seperate file for each controller function.
